changed the type for duration (int -> double)

+++ teamBWinterschool/src/entities/PrioritizedTask.java
@@ -4,5 +4,5 @@ import org.eclipse.app4mc.amalthea.model.Task;
 public class PrioritizedTask {
 	Task task;
 	int recurrence;
-	int duration;
+	double duration; // hi
 }
